requirements.txt
pytest
tqdm
pandas
spacy
torch
transformers
matplotlib
faiss-cpu
